fix compile errors, remove cyclic import
