+++ src/fiatExchanges/PaymentMethodSection.test.tsx
@@ -17,7 +17,6 @@ import {
   mockFiatConnectQuotesWithUnknownFees,
   mockProviders,
 } from 'test/values'
-import { CiCoCurrency } from 'src/utils/currencies'
 
 const mockStore = createMockStore({
   localCurrency: {
@@ -43,7 +42,7 @@ describe('PaymentMethodSection', () => {
     props = {
       paymentMethod: PaymentMethod.Card,
       // the below creates 4 quotes - 1 Ramp (card), 2 Moonpay (bank, card), 1 Simplex (card)
-      normalizedQuotes: normalizeQuotes(CICOFlow.CashIn, [], mockProviders, CiCoCurrency.CUSD),
+      normalizedQuotes: normalizeQuotes(CICOFlow.CashIn, [], mockProviders, CiCoCurrency.cUSD),
       setNoPaymentMethods: jest.fn(),
       flow: CICOFlow.CashIn,
       cryptoType: CiCoCurrency.cUSD,
@@ -67,7 +66,7 @@ describe('PaymentMethodSection', () => {
       CICOFlow.CashIn,
       [],
       [mockProviders[2]],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     const { queryByText, queryByTestId } = render(
       <Provider store={mockStore}>
@@ -84,7 +83,7 @@ describe('PaymentMethodSection', () => {
       CICOFlow.CashIn,
       [],
       [mockProviders[2]],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     jest.spyOn(props.normalizedQuotes[0], 'isProviderNew').mockReturnValue(true)
     const { getByText, getByTestId } = render(
@@ -167,7 +166,7 @@ describe('PaymentMethodSection', () => {
       CICOFlow.CashIn,
       mockFiatConnectQuotesWithUnknownFees,
       [],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     props.paymentMethod = PaymentMethod.Bank
     const { queryByText, queryByTestId } = render(
@@ -186,7 +185,7 @@ describe('PaymentMethodSection', () => {
       CICOFlow.CashIn,
       [mockFiatConnectQuotes[3]] as FiatConnectQuoteSuccess[],
       [],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     props.paymentMethod = PaymentMethod.Bank
     const { queryByTestId } = render(
@@ -206,7 +205,7 @@ describe('PaymentMethodSection', () => {
       CICOFlow.CashIn,
       [mockFiatConnectQuotes[1]] as FiatConnectQuoteSuccess[],
       [],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     props.paymentMethod = PaymentMethod.Bank
     const { queryByTestId } = render(
@@ -223,7 +222,7 @@ describe('PaymentMethodSection', () => {
   it.each([
     [
       PaymentMethod.Card as const,
-      normalizeQuotes(CICOFlow.CashIn, [], [mockProviders[2]], CiCoCurrency.CUSD),
+      normalizeQuotes(CICOFlow.CashIn, [], [mockProviders[2]], CiCoCurrency.cUSD),
       'card',
       'oneHour',
     ],
@@ -233,7 +232,7 @@ describe('PaymentMethodSection', () => {
         CICOFlow.CashIn,
         [mockFiatConnectQuotes[1]] as FiatConnectQuoteSuccess[],
         [],
-        CiCoCurrency.CUSD
+        CiCoCurrency.cUSD
       ),
       'bank',
       'numDays',
@@ -244,7 +243,7 @@ describe('PaymentMethodSection', () => {
         CICOFlow.CashIn,
         [mockFiatConnectQuotes[4]] as FiatConnectQuoteSuccess[],
         [],
-        CiCoCurrency.CUSD
+        CiCoCurrency.cUSD
       ),
       'mobileMoney',
       'lessThan24Hours',
