+++ src/fiatExchanges/quotes/normalizeQuotes.test.ts
@@ -6,13 +6,13 @@ import {
 } from 'src/fiatExchanges/quotes/normalizeQuotes'
 import { CICOFlow } from 'src/fiatExchanges/utils'
 import { TokenBalance } from 'src/tokens/slice'
+import { CiCoCurrency } from 'src/utils/currencies'
 import Logger from 'src/utils/Logger'
 import {
   mockFiatConnectQuotes,
   mockFiatConnectQuotesWithUnknownFees,
   mockProviders,
 } from 'test/values'
-import { CiCoCurrency } from 'src/utils/currencies'
 
 jest.mock('src/utils/Logger', () => ({
   __esModule: true,
@@ -29,7 +29,7 @@ describe('normalizeQuotes', () => {
       CICOFlow.CashIn,
       mockFiatConnectQuotes,
       mockProviders,
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     expect(
       normalizedQuotes.map((quote) => [
@@ -61,7 +61,7 @@ describe('normalizeQuotes', () => {
       CICOFlow.CashIn,
       mockFiatConnectQuotesWithUnknownFees,
       [],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     expect(
       normalizedQuotes.map((quote) => [
@@ -127,7 +127,7 @@ describe('normalizeExternalProviders', () => {
     const normalizedExternalQuotes = normalizeExternalProviders(
       CICOFlow.CashIn,
       [mockProviders[3]],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     expect(Logger.warn).toHaveBeenCalledWith(
       'NormalizeQuotes',
@@ -140,7 +140,7 @@ describe('normalizeExternalProviders', () => {
     const normalizedExternalQuotes = normalizeExternalProviders(
       CICOFlow.CashIn,
       [mockProviders[1]],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     expect(Logger.warn).not.toHaveBeenCalled()
     expect(normalizedExternalQuotes).toHaveLength(2)
@@ -150,7 +150,7 @@ describe('normalizeExternalProviders', () => {
     const normalizedExternalQuotes = normalizeExternalProviders(
       CICOFlow.CashIn,
       [mockProviders[0]],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     expect(Logger.warn).not.toHaveBeenCalled()
     expect(normalizedExternalQuotes).toHaveLength(1)
@@ -160,7 +160,7 @@ describe('normalizeExternalProviders', () => {
     const normalizedExternalQuotes = normalizeExternalProviders(
       CICOFlow.CashOut,
       [mockProviders[6]],
-      CiCoCurrency.CUSD
+      CiCoCurrency.cUSD
     )
     expect(Logger.warn).not.toHaveBeenCalled()
     expect(normalizedExternalQuotes).toHaveLength(2)
